// src/api-client/models/token-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EntityBasic } from './entity-basic';

/**
 * 
 * @export
 * @interface TokenResponse
 */
export interface TokenResponse {
    /**
     * 
     * @type {string}
     * @memberof TokenResponse
     */
    'accessToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof TokenResponse
     */
    'refreshToken'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TokenResponse
     */
    'authenticated'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TokenResponse
     */
    'email'?: string;
    /**
     * 
     * @type {Set<EntityBasic>}
     * @memberof TokenResponse
     */
    'roles'?: Set<EntityBasic>;
}